Fix attempt to use non-existent name
This theoretically also changes the meaning of the code from running
the callback always, to only running the callback when there wasn't an
error, which is what I'm assuming was meant, but in pratical terms,
attempting to use `data` in the case of an error would have raised a
`NameError` anyway.

`else` on a try block will only run if nothing was raised.