Stopped using hy for keybindings
keybindings are now a JSON format.
Somehow hy errored for me, and the cleanest solution seemed not to use hy at all
Thinking about it, the client was basically loading a new interpreter to optimize something that didn't need optimizing.
It didn't make it more readable/writable either, and it always was a lot of hassle.